mock-oidc/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
PyJWT==2.9.0
cryptography==42.0.5
python-multipart==0.0.20
